+++ docs/source/conf.py
@@ -28,6 +28,7 @@ extensions = [
     'sphinx.ext.mathjax',             # Render Math in html outputs
     'sphinx.ext.napoleon',            # Support Numpy-Styled docstrings
     'autoapi.extension',              # Automatically creating .rst files from docstrings
+    'sphinx_design',                  # Responsive Websites
     'myst_nb',                        # Support Notebooks and Markdown
 ]
 
@@ -54,7 +55,7 @@ autoapi_dirs = [
     '../../src/fairscoring',
 ]
 autoapi_root = "autoapi"
-autoapi_add_toctree_entry = True   # Automatically add
+autoapi_add_toctree_entry = False  # Automatically add
 autoapi_options = [
     'members',                     # Display children of an object
     'inherited-members',           # Display children of an object that have been inherited from a base class.
@@ -64,7 +65,7 @@ autoapi_options = [
     'show-inheritance',            # Display a list of base classes below the class signature.
     # 'show-inheritance-diagram',  # Display an inheritance diagram in generated class documentation.'
     'show-module-summary',         # Whether to include autosummary directives in generated module documentation.
-    # 'imported-members',            # Display objects imported from the same top level package or module.
+    'imported-members',            # Display objects imported from the same top level package or module.
 ]
 
 # Turn on for debugging only
@@ -100,7 +101,7 @@ nb_execution_mode = "off"
 myst_enable_extensions  = [
     "amsmath",
     # "attrs_inline",
-    # "colon_fence",
+    "colon_fence",
     # "deflist",
     "dollarmath",
     # "fieldlist",